pyproject.toml
[tool.poetry]
name = "homerp-finance"
version = "0.1.0"
description = "HomeRP Finance module"
authors = ["Beilak <beylak@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.9.4"
fastapi = "^0.99.0"
dependency-injector = "^4.41.0"
pydantic = "<2.0.0"
sqlalchemy = "^2.0.4"
uvicorn = "*"
sqlalchemy-utils = "*"
py-moneyed = "^3.0"
psycopg2-binary = "^2.9.5"
asyncpg = "^0.27.0"
python-keycloak = "*"
cryptography = "*"
secure = "^0.3.0"
aio-pika = "^9.1.3"
redis = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
